ktb_web4/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
huggingface-hub
sentence-transformers
faiss-cpu
rank_bm25
tf-keras
